// peage-pay-server/src/auth/graphql/auth-errors.graphql.ts
import { registerEnumType } from '@nestjs/graphql';

export enum AuthErrors {
  VERIFICATION_TOKEN_NOT_FOUND = 'VERIFICATION_TOKEN_NOT_FOUND',
  EMAIL_AUTH_NOT_FOUND = 'EMAIL_AUTH_NOT_FOUND',
  PHONE_AUTH_NOT_FOUND = 'PHONE_AUTH_NOT_FOUND',
  EMAIL_ALREADY_VERIFIED = 'EMAIL_ALREADY_VERIFIED',
  VERIFICATION_TOKEN_EXPIRED = 'VERIFICATION_TOKEN_EXPIRED',
  INVALID_VERIFICATION_TOKEN = 'INVALID_VERIFICATION_TOKEN',
  INVALID_EMAIL_OR_PASSWORD = 'INVALID_EMAIL_OR_PASSWORD',
  REFRESH_TOKEN_NOT_PROVIDED = 'REFRESH_TOKEN_NOT_PROVIDED',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  VERIFICATION_REQUEST_PENDING = 'VERIFICATION_REQUEST_PENDING',
  SIGN_IN_WITH_EMAIl_ATTEMPTS_EXCEEDED = 'SIGN_IN_WITH_EMAIl_ATTEMPTS_EXCEEDED',
  EMAIL_VERIFICATION_ATTEMPTS_EXCEEDED = 'EMAIL_VERIFICATION_ATTEMPTS_EXCEEDED',
  PASSWORD_RESET_ATTEMPTS_EXCEEDED = 'PASSWORD_RESET_ATTEMPTS_EXCEEDED',
}
registerEnumType(AuthErrors, {
  name: 'AuthErrors',
});